refactor : move gnb, header to global components
